# Increase harmonic processing speed with numba