--- a/rfpy/scripts/rfpy_harmonics.py
+++ b/rfpy/scripts/rfpy_harmonics.py
@@ -88,14 +88,6 @@ def get_harmonics_arguments(argv=None):
         "Default behaviour uses short key form (NET.STN) for the folder " +
         "names, regardless of the key type of the database."
     )
-    parser.add_argument(
-        "-UN", "--use-numba",
-        action="store_true",
-        dest="use_numba",
-        default=False,
-        help="Use numba jit on calculation [Default False]"
-    )
-
 
     # Event Selection Criteria
     TimeGroup = parser.add_argument_group(
@@ -217,6 +209,13 @@ def get_harmonics_arguments(argv=None):
         default=False,
         help="Set this option to save the Harmonics object " +
         "to a pickled file. [Default does not save object]")
+    HarmonicGroup.add_argument(
+        "--use-numba",
+        action="store_true",
+        dest="use_numba",
+        default=False,
+        help="Use numba jit to increase processing speed [Default does not use numba]"
+    )
 
     PlotGroup = parser.add_argument_group(
         title='Settings for plotting results',
@@ -510,8 +509,9 @@ def main():
 
         # Stack with or without dip
         if args.find_azim:
+            # Check if the process use numba
             if args.use_numba:
-                harmonics.dcomp_find_azim_numba(xmin=args.trange[0], xmax=args.trange[1])
+                harmonics.dcomp_find_azim(xmin=args.trange[0], xmax=args.trange[1], use_numba=True)
             else:
                 harmonics.dcomp_find_azim(xmin=args.trange[0], xmax=args.trange[1])